// utils/enums.ts
enum BASEBALL_TABLE_NAMES {
  baseballHomeNonDivision = "BASEBALL HOME NON DIVISION",
  baseballHomeDivision = "BASEBALL HOME DIVISION",
  baseballHomeNonDivision2022 = "BASEBALL HOME NON DIVISION 2022",
  baseballHomeDivision2022 = "BASEBALL HOME DIVISION 2022",
  baseballVisitorNonDivision = "BASEBALL VISITOR NON DIVISION",
  baseballVisitorDivision = "BASEBALL VISITOR DIVISION",
  baseballVisitorNonDivision2022 = "BASEBALL VISITOR NON DIVISION 2022",
  baseballVisitorDivision2022 = "BASEBALL VISITOR DIVISION 2022",
}

enum SPORT_NAME {
  BASEBALL = "BASEBALL",
  FOOTBALL = "FOOTBALL",
}

export { BASEBALL_TABLE_NAMES, SPORT_NAME };